requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
alembic>=1.10.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.24.0
gunicorn>=21.2.0
websockets>=11.0
requests>=2.31.0
bcrypt==4.0.1
tenacity>=8.2.0
python-multipart>=0.0.6
email-validator>=2.0.0
google-generativeai>=0.3.0
